requirements.txt
streamlit
google-generativeai
python-dotenv
pdfplumber
PyPDF2
